move NotificationItem component to folder
