+++ src/modules/stackables/components/notificationItem/NotificationItem.js
@@ -3,10 +3,10 @@
  */
 import { html, nothing } from 'lit-html';
 import { classMap } from 'lit-html/directives/class-map.js';
-import { $injector } from '../../../injection';
-import { LevelTypes } from '../../../store/notifications/notifications.action';
+import { $injector } from '../../../../injection';
+import { LevelTypes } from '../../../../store/notifications/notifications.action';
 import css from './notificationItem.css';
-import { MvuElement } from '../../MvuElement';
+import { MvuElement } from '../../../MvuElement';
 
 export const NOTIFICATION_AUTOCLOSE_TIME_NEVER = 0;
